src/requirements.txt
python-dotenv
requests
streamlit
langchain-openai
lanchain-core
lanchain-text-splitters